mkRespMod {MatrixModels}	R Documentation

Create a respModule object

Description

Create a respModule object, which could be from a derived class such as glmRespMod or nlsRespMod.

Usage

mkRespMod(fr, family = NULL, nlenv = NULL, nlmod = NULL)

Arguments

fr
a model frame, usually created by a call to model.frame.

family
an optional glm family object (glmRespMod objects only).

nlenv
an environment for evaluation of the nonlinear model, nlmod. (nlsRespMod objects only).

nlmod
the nonlinear model function, as a function call (nlsRespMod objects only).

Details

The internal representation of a statistical model based on a linear predictor expression is derived from a formula expression and a data argument, possibly supplemented with a family object and/or a nonlinear model expression. The steps to obtain this representation usually involve calls to model.frame and to model.matrix or model.Matrix, which encapsulate important parts of this process. This function encapsulates other operations related to weights and offsets and to the model family to create a respModule object.

Value

an object of a class inheriting from respModule.

See Also

The respModule class description.

Examples

  ## see  help("glpModel-class")

[Package MatrixModels version 0.5-3 Index]