resid-et-al {MatrixModels}	R Documentation

Aliases for Model Extractors

Description

Aliases for model extractors; it is an old S and R tradition to have aliases for these three model extractor functions:

resid()

equivalent to residuals().

fitted.values()

equivalent to fitted().

coefficients()

equivalent to coef().

We provide S4 generics and methods for these.

Methods

resid

signature(object = "ANY"): return the residuals; this is a rarely used alias for residuals().

fitted.values

signature(object = "ANY"): return the fitted values; this is a rarely used alias for fitted().

coefficients

signature(object = "ANY"): return the coefficients of a model; this is a rarely used alias for coef().

See Also

residuals; Methods for general information about formal (S4) methods.

[Package MatrixModels version 0.5-3 Index]