Model-class {MatrixModels}	R Documentation

Mother Class "Model" of all S4 Models

Description

Class "Model" is meant to be the mother class of all (S4) model classes. As some useful methods are already defined for "Model" objects, derived classes inherit those “for free”.

Objects from the Class

A virtual Class: No objects may be created from it.

Slots

call:

the call which generated the model.

fitProps:

a list; must be named, i.e., have unique names, but can be empty.

When the main object is a fitted model, the list will typically have components such as iter (non-negative integer) and convergenece (logical typically).

Methods

formula

signature(x = "Model"): extract the model formula - if there is one, or NULL.

update

signature(object = "Model"): Update the model with a new formula, new data, ...... etc. This semantically equivalent (and as R function almost identical) to the standard update (package stats).

See Also

the glpModel class in package MatrixModels which extends this class.

Examples

showClass("Model")

[Package MatrixModels version 0.5-0 Index]