glpModel-class {MatrixModels}	R Documentation

Class "glpModel" of General Linear Prediction Models

Description

The class "glpModel" conceptually contains a very large class of “General Linear Prediction Models”.

Its resp slot (of class "respModule") may model linear, non-linear, generalized linear and non-linear generalized response models.

Objects from the Class

Objects can be created by calls of the form new("glpModel", ...), but typically rather are returned by our modeling functions, e.g., the (experimental, hence currently hidden) glm4().

Slots

resp:

a "respModule" object.

pred:

a "predModule" object.

Extends

Class "Model", directly.

Methods

coef

signature(object = "glpModel"): extract the coefficient vector β from the object.

fitted

signature(object = "glpModel"): fitted values; there may be several types, corresponding to the residuals, see there (below).

residuals

signature(object = "glpModel"): residuals, depending on the type of the model, there are several types of residuals and correspondingly residuals, see residuals.glm from the stats package.

See Also

glm4() returns fitted glpModel objects.

The constituents of this class are respModule and predModule, both of which have several sub classes.

Examples

showClass("glpModel")

## Use   example(glm4)  or see  help(glm4) for many more examples.

[Package MatrixModels version 0.5-0 Index]