updateMu {MatrixModels}	R Documentation

Update 'mu', the Fitted Mean Response

Description

Updates the mean vector μ given the linear predictor γ. Evaluate the residuals and the weighted sum of squared residuals.

Usage

updateMu(respM, gamma, ...)

Arguments

respM
a response module, see the respModule class.

gamma
the value of the linear predictor before adding the offset

...
potentially further arguments used in methods; not used currently.

Details

Note that the offset is added to the linear predictor before calculating mu.

The sqrtXwt matrix can be updated but the sqrtrwt should not be in that the weighted sum of squared residuals should be calculated relative to fixed weights. Reweighting is done in a separate call.

Value

updated respM

Methods

signature(respM = "glmRespMod", gamma = "numeric")

..

signature(respM = "nglmRespMod", gamma = "numeric")

..

signature(respM = "nlsRespMod", gamma = "numeric")

..

signature(respM = "respModule", gamma = "numeric")

..

See Also

The respModule class (and specific subclasses); glm4.

Examples

## TODO

[Package MatrixModels version 0.5-0 Index]